fix(adapters): address review feedback for NIU-589 saga templates
- Move _slugify to tyr/domain/utils.py (was tyr/api/tracker.py) — fixes
  layer violation where domain service and adapter imported from API layer
- Type _dispatch_template_phase tpl_phase as TemplatePhase instead of
  object; remove # type: ignore[union-attr]
- Remove _BUNDLED_TEMPLATES_DIR backward-compat alias and template re-exports
  from event_trigger.__all__; update all test imports to use
  BUNDLED_TEMPLATES_DIR directly from tyr.domain.templates
- Extract InMemorySagaRepository, StubVolundrPort, StubVolundrFactory into
  tests/test_tyr/stubs.py; both test files import from shared location
- advance_phase: upgrade untracked-saga log from DEBUG to WARNING with
  clear message about restart limitation; sync in-memory entry to GATED
  after _activate_phase gates a phase so status is not stale
